requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bootstrap4==0.1.0
certifi==2022.5.18.1
confusable-homoglyphs==3.2.0
distlib==0.3.4
Django==4.0.5
django-admin-honeypot==1.1.0
django-bootstrap4==22.3
django-crispy-forms==1.14.0
django-registration==3.3
django-session-timeout==0.1.0
filelock==3.7.0
Pillow==9.1.1
pipenv==2022.5.2
platformdirs==2.5.2
python-decouple==3.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
virtualenv==20.14.1
virtualenv-clone==0.5.7
